+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/manage/page.tsx
@@ -1,10 +1,8 @@
 import { serverTranslation } from "@i18n";
-import { getTemplateWithAssociatedUsers, checkIfClosed } from "@lib/templates";
+import { checkIfClosed } from "@lib/templates";
 import { authorization } from "@lib/privileges";
-import { getUsers } from "@lib/users";
-import { ManageForm } from "./ManageForm";
+import { ManageForm } from "./components/ManageForm";
 import { Metadata } from "next";
-import { headers } from "next/headers";
 import { AuthenticatedPage } from "@lib/pages/auth";
 
 export async function generateMetadata(props: {
@@ -28,10 +26,6 @@ export default AuthenticatedPage(async (props: { params: Promise<{ id: string }>
 
   let closedDetails;
 
-  const manageAllForms = await authorization
-    .canManageAllForms()
-    .then(() => true)
-    .catch(() => false);
   const canSetClosingDate =
     id !== "0000" ||
     (await authorization
@@ -39,53 +33,10 @@ export default AuthenticatedPage(async (props: { params: Promise<{ id: string }>
       .then(() => true)
       .catch(() => false));
 
-  const nonce = (await headers()).get("x-nonce");
-
   if (canSetClosingDate) {
     const closedData = await checkIfClosed(id);
     closedDetails = closedData?.closedDetails;
   }
 
-  // ⚠️ This is the main entry point for the manage form page. Code beyond this if block
-  // is only reached if the user is an "admin" (can manage all forms)
-  if (!manageAllForms || id === "0000") {
-    return (
-      <ManageForm
-        nonce={nonce}
-        id={id}
-        canManageAllForms={false}
-        canSetClosingDate={canSetClosingDate}
-        closedDetails={closedDetails}
-      />
-    );
-  }
-
-  const templateWithAssociatedUsers = await getTemplateWithAssociatedUsers(id);
-
-  if (!templateWithAssociatedUsers) {
-    throw new Error("Template not found");
-  }
-
-  const allUsers = await getUsers().then((users) =>
-    users.map((user) => ({
-      id: user.id,
-      name: user.name || "",
-      email: user.email || "",
-    }))
-  );
-
-  return (
-    <>
-      <ManageForm
-        nonce={nonce}
-        id={id}
-        canManageAllForms={manageAllForms}
-        canSetClosingDate={canSetClosingDate}
-        formRecord={templateWithAssociatedUsers.formRecord}
-        usersAssignedToFormRecord={templateWithAssociatedUsers.users}
-        allUsers={allUsers}
-        closedDetails={closedDetails}
-      />
-    </>
-  );
+  return <ManageForm id={id} canSetClosingDate={canSetClosingDate} closedDetails={closedDetails} />;
 });
